+++ src/aui-layout/HISTORY.md
@@ -3,7 +3,7 @@
 > Documentation and test modifications are not included in this changelog. For more details, see [full commit history](https://github.com/liferay/alloy-ui/commits/master/src/aui-layout).
 
 ## @VERSION@
-
+* [AUI-1727](https://issues.liferay.com/browse/AUI-1727) Add keyboard navigation to layout builder
 * [AUI-1711](https://issues.liferay.com/browse/AUI-1711) Improve interaction of adding new rows
 * [AUI-1745](https://issues.liferay.com/browse/AUI-1745) Investigate and fix performance problems on Layout Builder
 * [AUI-1726](https://issues.liferay.com/browse/AUI-1726) Position buttons to add/remove columns according to wireframe